release: v3.6 opencode final - 日期安全/仅删除确认/解析提速/服务器时钟
- fix: 无日期输入时新增账目强制用客户端当前时间，不再采信模型从历史带出的旧日期（修复时间停到 8.1）
- fix: 修改账目时用户未提时间则不挪动日期，保留原日期
- fix: 仅删除操作需要确认；批量新增与修改直接执行（按产品准则）
- perf: 删改上下文 200->30 条、历史消息截断 4000->1200 字符、打字机速度 3 倍
- feat: 云函数 /health 返回 serverTime；AI 会员页新增服务器在线状态与服务器时间显示
- test: 164 个单元测试全部通过

+++ app/src/main/java/com/expense/tracker/ui/template/TemplateCard.kt
@@ -20,7 +20,7 @@ import com.expense.tracker.ui.theme.AppColors
 fun TemplateCard(
     selectedCategoryId: String,
     onSelectCategory: (String) -> Unit,
-    onSubmit: (Double) -> Unit,
+    onSubmit: (Long) -> Unit,
     onDoubleClickCategory: (String) -> Unit = {},
     modifier: Modifier = Modifier,
 ) {
